mm: madvise MADV_USERFAULT: prepare vm_flags to allow more than 32bits
We run out of 32bits in vm_flags, noop change for 64bit archs.

+++ include/linux/mm_types.h
@@ -217,7 +217,7 @@ struct page_frag {
 #endif
 };
 
-typedef unsigned long __nocast vm_flags_t;
+typedef unsigned long long __nocast vm_flags_t;
 
 /*
  * A region containing a mapping of a non-memory backed file under NOMMU
